+++ R/rdb.R
@@ -28,6 +28,8 @@
 #' package \pkg{jsonlite}) is applied to generate the json object.
 #' @param mask Character string (default \code{NULL}). DBnomics code of one or
 #' several masks in the specified provider and dataset.
+#' @param query Character string (default \code{NULL}). A query to
+#' filter/select series from a provider's dataset.
 #' @param filters List (default \code{NULL}). This argument must be a named
 #' list for one filter because the function \code{toJSON} of the package \pkg{jsonlite}
 #' is used before sending the request to the server. For multiple filters,
@@ -44,58 +46,65 @@
 #' \dontrun{
 #' ## By ids
 #' # Fetch one series from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
-#' df1 <- rdb(ids = 'AMECO/ZUTN/EA19.1.0.0.0.ZUTN')
+#' df1 <- rdb(ids = "AMECO/ZUTN/EA19.1.0.0.0.ZUTN")
 #' # or when no argument names are given (provider_code -> ids)
-#' df1 <- rdb('AMECO/ZUTN/EA19.1.0.0.0.ZUTN')
+#' df1 <- rdb("AMECO/ZUTN/EA19.1.0.0.0.ZUTN")
 #'
 #' # Fetch two series from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
-#' df2 <- rdb(ids = c('AMECO/ZUTN/EA19.1.0.0.0.ZUTN', 'AMECO/ZUTN/DNK.1.0.0.0.ZUTN'))
+#' df2 <- rdb(ids = c("AMECO/ZUTN/EA19.1.0.0.0.ZUTN", "AMECO/ZUTN/DNK.1.0.0.0.ZUTN"))
 #'
 #' # Fetch two series from different datasets of different providers :
-#' df3 <- rdb(ids = c('AMECO/ZUTN/EA19.1.0.0.0.ZUTN', 'IMF/CPI/A.AT.PCPIT_IX'))
+#' df3 <- rdb(ids = c("AMECO/ZUTN/EA19.1.0.0.0.ZUTN", "IMF/CPI/A.AT.PCPIT_IX"))
 #'
 #'
 #' ## By dimensions
 #' # Fetch one value of one dimension from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
-#' df1 <- rdb('AMECO', 'ZUTN', dimensions = list(geo = "ea12"))
+#' df1 <- rdb("AMECO", "ZUTN", dimensions = list(geo = "ea12"))
 #' # or
-#' df1 <- rdb('AMECO', 'ZUTN', dimensions = '{"geo": ["ea12"]}')
+#' df1 <- rdb("AMECO", "ZUTN", dimensions = '{"geo": ["ea12"]}')
 #'
 #' # Fetch two values of one dimension from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
-#' df2 <- rdb('AMECO', 'ZUTN', dimensions = list(geo = c("ea12", "dnk")))
+#' df2 <- rdb("AMECO", "ZUTN", dimensions = list(geo = c("ea12", "dnk")))
 #' # or
-#' df2 <- rdb('AMECO', 'ZUTN', dimensions = '{"geo": ["ea12", "dnk"]}')
+#' df2 <- rdb("AMECO", "ZUTN", dimensions = '{"geo": ["ea12", "dnk"]}')
 #'
 #' # Fetch several values of several dimensions from dataset 'Doing business' (DB) of World Bank :
 #' dim <- list(
 #'   country = c("DZ", "PE"),
 #'   indicator = c("ENF.CONT.COEN.COST.ZS", "IC.REG.COST.PC.FE.ZS")
 #' )
-#' df3 <- rdb('WB', 'DB', dimensions = dim)
+#' df3 <- rdb("WB", "DB", dimensions = dim)
 #' # or
 #' dim <- paste0(
 #'   '{"country": ["DZ", "PE"],',
 #'   '"indicator": ["ENF.CONT.COEN.COST.ZS", "IC.REG.COST.PC.FE.ZS"]}'
 #' )
-#' df3 <- rdb('WB', 'DB', dimensions = dim)
+#' df3 <- rdb("WB", "DB", dimensions = dim)
 #'
 #'
 #' ## By mask
 #' # Fetch one series from dataset 'Consumer Price Index' (CPI) of IMF :
-#' df1 <- rdb('IMF', 'CPI', mask = 'M.DE.PCPIEC_WT')
+#' df1 <- rdb("IMF", "CPI", mask = "M.DE.PCPIEC_WT")
 #' # or when no argument names are given except provider_code and dataset_code (ids -> mask)
-#' df1 <- rdb('IMF', 'CPI', 'M.DE.PCPIEC_WT')
+#' df1 <- rdb("IMF", "CPI", "M.DE.PCPIEC_WT")
 #'
 #' # Fetch two series from dataset 'Consumer Price Index' (CPI) of IMF :
-#' df2 <- rdb('IMF', 'CPI', mask = 'M.DE+FR.PCPIEC_WT')
+#' df2 <- rdb("IMF", "CPI", mask = "M.DE+FR.PCPIEC_WT")
 #'
 #' # Fetch all series along one dimension from dataset 'Consumer Price Index' (CPI) of IMF :
-#' df3 <- rdb('IMF', 'CPI', mask = 'M..PCPIEC_WT')
+#' df3 <- rdb("IMF", "CPI", mask = "M..PCPIEC_WT")
 #'
 #' # Fetch series along multiple dimensions from dataset 'Consumer Price Index' (CPI) of IMF :
-#' df4 <- rdb('IMF', 'CPI', mask = 'M..PCPIEC_IX+PCPIA_IX')
+#' df4 <- rdb("IMF", "CPI", mask = "M..PCPIEC_IX+PCPIA_IX")
 #'
 #'
+#' ## By query
+#' # Fetch one series from dataset 'WEO by countries' (WEO) from IMF :
+#' df1 <- rdb("IMF", "WEO", query = "France current account balance percent")
+#' # Fetch series from dataset 'WEO by countries' (WEO) from IMF :
+#' df2 <- rdb("IMF", "WEO", query = "current account balance percent")
+#' 
+#' 
 #' ## Use a specific proxy to fetch the data
 #' # Fetch one series from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
 #' h <- list(
@@ -105,18 +114,18 @@
 #'   proxypassword = "<password>"
 #' )
 #' options(rdbnomics.curl_config = h)
-#' df1 <- rdb(ids = 'AMECO/ZUTN/EA19.1.0.0.0.ZUTN')
+#' df1 <- rdb(ids = "AMECO/ZUTN/EA19.1.0.0.0.ZUTN")
 #' # or to use once
 #' options(rdbnomics.curl_config = NULL)
-#' df1 <- rdb(ids = 'AMECO/ZUTN/EA19.1.0.0.0.ZUTN', curl_config = h)
+#' df1 <- rdb(ids = "AMECO/ZUTN/EA19.1.0.0.0.ZUTN", curl_config = h)
 #'
 #'
 #' ## Use R default connection to avoid a proxy failure (in some cases)
 #' # Fetch one series from dataset 'Unemployment rate' (ZUTN) of AMECO provider :
 #' options(rdbnomics.use_readLines = TRUE)
-#' df1 <- rdb(ids = 'AMECO/ZUTN/EA19.1.0.0.0.ZUTN')
+#' df1 <- rdb(ids = "AMECO/ZUTN/EA19.1.0.0.0.ZUTN")
 #' # or to use once
-#' df1 <- rdb(ids = 'AMECO/ZUTN/EA19.1.0.0.0.ZUTN', use_readLines = TRUE)
+#' df1 <- rdb(ids = "AMECO/ZUTN/EA19.1.0.0.0.ZUTN", use_readLines = TRUE)
 #' 
 #' 
 #' ## Apply filter(s) to the series
@@ -148,7 +157,7 @@
 #' @export
 rdb <- function(
   provider_code = NULL, dataset_code = NULL,
-  ids = NULL, dimensions = NULL, mask = NULL,
+  ids = NULL, dimensions = NULL, mask = NULL, query = NULL,
   filters = getOption("rdbnomics.filters"),
   verbose = getOption("rdbnomics.verbose_warning"),
   ...
@@ -187,6 +196,9 @@ rdb <- function(
 
   ids_null <- is.null(ids)
   ids_not_null <- !ids_null
+
+  query_null <- is.null(query)
+  query_not_null <- !query_null
 
   # provider_code is considered as ids in some cases
   if (
@@ -345,8 +357,51 @@ rdb <- function(
     return(rdb_by_api_link(api_link = link, filters = filters, ...))
   }
 
+  # By query
+  if (query_not_null) {
+    if (provider_code_null | dataset_code_null) {
+      stop(
+        paste0(
+          "When you filter with a 'query', you must specifiy 'provider_code' ",
+          "and 'dataset_code' as arguments of the function."
+        ),
+        call. = FALSE
+      )
+    }
+
+    check_argument(query, "character", not_null = FALSE)
+
+    if (verbose) {
+      if (query == "") {
+        warning(
+          paste0(
+            "Your 'query' is empty, the entire dataset ",
+            provider_code, "/", dataset_code,
+            " will be downloaded. It can be long !"
+          )
+        )
+      }
+    }
+
+    if (api_version == 22) {
+      link <- paste0(
+        api_base_url, "/", provider_code, "/", dataset_code,
+        "?q=", utils::URLencode(query),
+        ifelse(metadata, "&", paste0("&metadata=", as.numeric(metadata), "&")),
+        "observations=1"
+      )
+    } else {
+      stop(
+        paste0("Don't know what to do for API version ", api_version, "."),
+        call. = FALSE
+      )
+    }
+
+    return(rdb_by_api_link(api_link = link, filters = filters, ...))
+  }
+
   stop(
-    "Please provide correct 'dimensions', 'mask', 'ids' or 'filters'.",
+    "Please provide correct 'dimensions', 'mask', 'ids', 'query' or 'filters'.",
     call. = FALSE
   )
 }