+++ R/rdb_by_api_link.R
@@ -36,17 +36,17 @@
 #' # Fetch two series from different datasets of different providers :
 #' df1 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/',
-#'     'series?observations=1&series_ids=AMECO/ZUTN/EA19.1.0.0.0.ZUTN,IMF/CPI/A.AT.PCPIT_IX'
+#'     "https://api.db.nomics.world/v22/",
+#'     "series?observations=1&series_ids=AMECO/ZUTN/EA19.1.0.0.0.ZUTN,IMF/CPI/A.AT.PCPIT_IX"
 #'   )
 #' )
 #' 
 #' # Fetch one series from the dataset 'Doing Business' of WB provider :
 #' df2 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22',
-#'     'indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business',
-#'     '&observations=1&format=json&align_periods=1&offset=0&facets=0'
+#'     "https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22",
+#'     "indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business",
+#'     "&observations=1&format=json&align_periods=1&offset=0&facets=0"
 #'   )
 #' )
 #' 
@@ -62,17 +62,17 @@
 #' options(rdbnomics.curl_config = h)
 #' df2 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22',
-#'     'indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business',
-#'     '&observations=1&format=json&align_periods=1&offset=0&facets=0'
+#'     "https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22",
+#'     "indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business",
+#'     "&observations=1&format=json&align_periods=1&offset=0&facets=0"
 #'   )
 #' )
 #' # or to use once
 #' df2 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22',
-#'     'indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business',
-#'     '&observations=1&format=json&align_periods=1&offset=0&facets=0'
+#'     "https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22",
+#'     "indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business",
+#'     "&observations=1&format=json&align_periods=1&offset=0&facets=0"
 #'   ),
 #'   curl_config = h
 #' )
@@ -83,17 +83,17 @@
 #' options(rdbnomics.use_readLines = TRUE)
 #' df2 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22',
-#'     'indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business',
-#'     '&observations=1&format=json&align_periods=1&offset=0&facets=0'
+#'     "https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22",
+#'     "indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business",
+#'     "&observations=1&format=json&align_periods=1&offset=0&facets=0"
 #'   )
 #' )
 #' # or to use once
 #' df2 <- rdb_by_api_link(
 #'   paste0(
-#'     'https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22',
-#'     'indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business',
-#'     '&observations=1&format=json&align_periods=1&offset=0&facets=0'
+#'     "https://api.db.nomics.world/v22/series/WB/DB?dimensions=%7B%22",
+#'     "indicator%22%3A%5B%22IC.REG.PROC.FE.NO%22%5D%7D&q=Doing%20Business",
+#'     "&observations=1&format=json&align_periods=1&offset=0&facets=0"
 #'   ),
 #'   use_readLines = TRUE
 #' )
@@ -102,7 +102,7 @@
 #' ## Apply filter(s) to the series
 #' # One filter
 #' df3 <- rdb_by_api_link(
-#'   'https://api.db.nomics.world/v22/series/IMF/WEO/ABW.BCA?observations=1',
+#'   "https://api.db.nomics.world/v22/series/IMF/WEO/ABW.BCA?observations=1",
 #'   filters = list(
 #'     code = "interpolate",
 #'     parameters = list(frequency = "daily", method = "spline")
@@ -111,7 +111,7 @@
 #' 
 #' # Two filters
 #' df3 <- rdb_by_api_link(
-#'   'https://api.db.nomics.world/v22/series/IMF/WEO/ABW.BCA?observations=1',
+#'   "https://api.db.nomics.world/v22/series/IMF/WEO/ABW.BCA?observations=1",
 #'   filters = list(
 #'     list(
 #'       code = "interpolate",
@@ -486,6 +486,12 @@ rdb_by_api_link <- function(
     # Add boolean to distinct be filtered and non-filtered series
     DBdata[, filtered := FALSE]
     DBlist[, filtered := TRUE]
+
+    # For R 3.1
+    diff_cols <- setdiff(colnames(DBlist), colnames(DBdata))
+    if (length(diff_cols) > 0) {
+      DBdata[, period_middle_day := as.Date(NA)]
+    }
 
     DBdata <- list(DBdata, DBlist)
     DBdata <- data.table::rbindlist(DBdata, use.names = TRUE, fill = TRUE)