+++ R/rdbnomics.R
@@ -12,7 +12,8 @@ NULL
 if (getRversion() >= "2.15.1") {
   vars <- c(
     ".", ":=", "value", "dotI", "period", "period_start_day", "series_code",
-    "filtered", "original_period", "series_name", "original_value"
+    "filtered", "original_period", "series_name", "original_value",
+    "period_middle_day"
   )
   utils::globalVariables(unique(vars))
 }